ref #44 moving metrics forward  logic to a custom scrapper
