+++ src/exporter/exporter.go
@@ -8,75 +8,18 @@ import (
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
-	"regexp"
-	"strings"
 
 	"github.com/NYTimes/gziphandler"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/simelo/rextporter/src/config"
+	"github.com/simelo/rextporter/src/scrapper"
 	"github.com/simelo/rextporter/src/util"
 	log "github.com/sirupsen/logrus"
 )
 
-func findMetricsName(metricsData string) (metricsNames []string) {
-	rex := regexp.MustCompile(`# TYPE [a-zA-Z_:][a-zA-Z0-9_:]*`)
-	metricsNameLines := rex.FindAllString(metricsData, -1)
-	metricsNames = make([]string, len(metricsNameLines))
-	for idx, metricsNameLine := range metricsNameLines {
-		metricsNameLineColumns := strings.Split(metricsNameLine, " ")
-		// FIXME(denissacostaq@gmail.com): be careful indexing here
-		metricsNames[idx] = metricsNameLineColumns[2]
-	}
-	return metricsNames
-}
-
-func appendPrefixForMetrics(prefix string, metricsData string) ([]byte, error) {
-	metricsName := findMetricsName(metricsData)
-	for _, metricName := range metricsName {
-		repl := strings.NewReplacer(
-			"# HELP "+metricName+" ", "# HELP "+prefix+"_"+metricName+" ",
-			"# TYPE "+metricName+" ", "# TYPE "+prefix+"_"+metricName+" ",
-		)
-		metricsData = repl.Replace(metricsData)
-		metricsData = strings.Replace(metricsData, "\n"+metricName, "\n"+prefix+"_"+metricName, -1)
-	}
-	if len(metricsName) == 0 {
-		err := fmt.Errorf("data from %s not appear to be from a metrics(trough prometheus instrumentation) endpoint", string(prefix))
-		log.WithError(err).Errorln("append prefix error, content ignored")
-	}
-	return []byte(metricsData), nil
-}
-
-func exposedMetricsMiddleware(metricsMiddleware []MetricMiddleware, promHandler http.Handler) http.Handler {
+func exposedMetricsMiddleware(fs scrapper.Scrapper, promHandler http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		getCustomData := func() (data []byte, err error) {
-			recorder := httptest.NewRecorder()
-			for _, mm := range metricsMiddleware {
-				if exposedMetricsData, err := mm.client.GetData(); err != nil {
-					log.WithError(err).Error("error getting metrics from service " + mm.client.ServiceName)
-				} else {
-					if prefixed, err := appendPrefixForMetrics(mm.client.ServiceName, string(exposedMetricsData)); err == nil {
-						if count, err := recorder.Write(prefixed); err != nil || count != len(prefixed) {
-							if err != nil {
-								log.WithError(err).Errorln("error writing prefixed content")
-							}
-							if count != len(prefixed) {
-								log.WithFields(log.Fields{
-									"wrote":    count,
-									"required": len(prefixed),
-								}).Errorln("no enough content wrote")
-							}
-						}
-					}
-				}
-			}
-			if data, err = ioutil.ReadAll(recorder.Body); err != nil {
-				log.WithError(err).Errorln("can not read recorded custom data")
-				return nil, err
-			}
-			return data, nil
-		}
 		getDefaultData := func() (data []byte, err error) {
 			generalScopeErr := "error reding default data"
 			recorder := httptest.NewRecorder()
@@ -105,10 +48,17 @@ func exposedMetricsMiddleware(metricsMiddleware []MetricMiddleware, promHandler 
 		} else {
 			allData = append(allData, defaultData...)
 		}
-		if customData, err := getCustomData(); err != nil {
+		var iMetrics interface{}
+		var err error
+		if iMetrics, err = fs.GetMetric(); err != nil {
 			log.WithError(err).Errorln("error getting custom data")
 		} else {
-			allData = append(allData, customData...)
+			customData, okCustomData := iMetrics.([]byte)
+			if okCustomData {
+				allData = append(allData, customData...)
+			} else {
+				log.WithError(err).Errorln("error getting custom data")
+			}
 		}
 		w.Header().Set("Content-Type", "text/plain")
 		if allData == nil {
